lhsm: assume ARCHIVE_RUNNING on HS_EXISTS
llapi_hsm_state_get() doesn't return the current actions anymore.
Setting a file as MODIFIED if it is only has HS_EXISTS
is wrong and will trigger duplicate archives.

This patch changes the behavior so that when a file is only
in the exists state, we assume it is currently archiving.

What could go wrong? If archiving is canceled, the 'archiving'
status could stay in the DB. Occasional cleanups might
be needed (manual archiving->new) until this is fixed in a better way.

Change-Id: Ia0d262abc722eb562fd5447bc6db4c3d7ba48bce